Demonstrate bug with `,` command jumping to the wrong line in blame view
@krobelus observed in https://github.com/jonas/tig/pull/1370 that this
behavior, fixed in ca0809dbd20, regressed again in 2280734104f.

This bug is similar to, but not identical with,
https://github.com/jonas/tig/issues/1369.

The easiest way to observe the bug is to run `tig blame +28 tig-2.5.12
-- src/diff.c` and press `,` AKA `:parent`. The problem is that the view
stays on line 28 instead of jumping to the corresponding line.

The test is almost an exact copy of the test by @krobelus
from #1370.